requirements.txt
pyyaml
numpy 
pandas 
matplotlib 
plotly
scikit-learn
xgboost
transformers
datasets
umap-learn
tensorflow